Implement and run the transformation of events into the unified format #18
We have developed an utility for write a RDD to Parquet. So we can test the utility that  we have  made we have created
a Scala test.

The test read from a XDR csv with near 15 million rows and use an Event Scheme for serialize an Avro file into Parquet.
We have made writes in Parquet without Avro so we can compare sizes and performance.

Test has been executed over a local filesystem and distributed HDFS too.

Next steps should be discussed between team members.